req 11 ++

+++ controllers/postController.js
@@ -38,8 +38,17 @@ routerPost.post('/', validateToken, validatePost, async (req, res) => {
   res.status(201).json(newPost);
 });
 
-routerPost.delete('/me', validateToken, validateToken, async (req, res) => {
-  const { id } = req.user;
+routerPost.delete('/:id', validateToken, async (req, res) => {
+  const { id } = req.params;
+  const { user } = req;
+  console.log('aquiiiiiiii2', user.id);
+
+  const post = await BlogPosts.findByPk(
+    req.params.id, { include: { model: User, as: 'user' } },
+  );
+  console.log('aquiiiiiiii', post);
+  if (!post) return res.status(404).json({ message: 'Post não existe' });
+  if (post.dataValues.userId !== user.id) return res.status(401).json({ message: 'Usuário não autorizado' });
 
   BlogPosts.destroy({ where: { id } });
   res.status(204).json({});
